Refactor shared memory access in consumer.c
When both programs are run at the same time, the consumer process may attempt to open the shared memory before the producer process could allocate it. As a result, the producer will be waiting while the consumer process faces a segmentation fault. Upon running the the two programs again, the terminal may not output the initial messages from the producer. This is because the shared memory has not been deallocated in the directory /dev/shm. To fix this, the consumer is forced into a loop until the shared memory has been allocated first.